гененация массива объектов. добавление модульности
