+++ js/main.js
@@ -1,135 +1,1 @@
-/*Структура объекта:
-id, число — идентификатор опубликованной фотографии. Это число от 1 до 25. Идентификаторы не должны повторяться.
-url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25. Адреса картинок не должны повторяться.
-description, строка — описание фотографии. Описание придумайте самостоятельно.
-likes, число — количество лайков, поставленных фотографии. Случайное число от 15 до 200.
-comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии. Количество комментариев к
-каждой фотографии вы определяете на своё усмотрение. Все комментарии генерируются случайным образом. Пример описания объекта с комментарием:
-  {
-    id: 135,
-    avatar: 'img/avatar-6.svg',
-    message: 'В целом всё неплохо. Но не всё.',
-    name: 'Артём',
-  }
-  У каждого комментария есть идентификатор — id — любое число. Идентификаторы не должны повторяться.
-  Поле avatar — это строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg. Аватарки подготовлены в директории img.
-  // Для формирования текста комментария — message — вам необходимо взять одно или два случайных предложения из представленных ниже:
-  //   Всё отлично!
-  //   В целом всё неплохо. Но не всё.
-  //   Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
-  //   Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
-  //   Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
-  //   Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!
-  // Имена авторов также должны быть случайными. Набор имён для комментаторов составьте сами. Подставляйте случайное имя в поле name.*/
-
-const AMOUNT_PHOTO = 25;
-const MAX_AMOUNT_COMMENTS_IN_OBJECT = 10;
-const MIN_AMOUNT_LIKES = 15;
-const MAX_AMOUNT_LIKES = 200;
-const AMOUNT_AVIALABEL_AVATARS = 6;
-
-const DESCRIPTIONS = [
-  'описание фотографии один',
-  'описание фотографии два',
-  'описание фотографии три',
-  'описание фотографии четыре',
-  'описание фотографии пять',
-  'описание фотографии шесть',
-  'описание фотографии семь',
-  'описание фотографии восемь',
-  'описание фотографии девять',
-  'описание фотографии десять',
-  'описание фотографии одиннадцать',
-  'описание фотографии двенадцать',
-  'описание фотографии тринадцать',
-  'описание фотографии четырнадцать',
-  'описание фотографии пятнадцать',
-  'описание фотографии шестнадцать',
-  'описание фотографии семнадцать',
-  'описание фотографии восемнадцать',
-  'описание фотографии девятнадцать',
-  'описание фотографии двадцать',
-  'описание фотографии двадцать один',
-  'описание фотографии двадцать два',
-  'описание фотографии двадцать три',
-  'описание фотографии двадцать четыре',
-  'описание фотографии двадцать пять'
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Люпита',
-  'Вашингтон',
-  'да Марья',
-  'Верон',
-  'Мирабелла',
-  'Вальц',
-  'Онопко',
-  'Топольницкая',
-  'Нионго',
-  'Ирвинг'
-];
-
-//Функция генерации уникальных идентификаторов
-function createIdGenerator (start) {
-  let lastGeneratedId = start;
-
-  return function () {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-}
-
-const generateObjectId = createIdGenerator(0);
-const generateCommentId = createIdGenerator(AMOUNT_PHOTO);
-
-// Функция генерации для получения уникальных целых чисел из указанного диапазона
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-};
-
-// Функция для получения случайного элемента массива, которая
-// принимает в качестве параметра один массив, а
-// возвращает его случайное значение.
-const arrayRandElement = (array) => array[getRandomInteger(0, array.length - 1)];
-
-// Функция-генератор комментария
-const generateCommentObject = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AMOUNT_AVIALABEL_AVATARS)}.svg`,
-  message: Array.from({length: getRandomInteger(1, 2)}, () => arrayRandElement(MESSAGES)).join(' '),
-  name: arrayRandElement(NAMES)
-});
-
-// Функция генератор объекта
-const generateObject = (objectId) => ({
-  id: objectId,
-  url: `photos/${objectId}.jpg`,
-  description: arrayRandElement(DESCRIPTIONS),
-  likes: getRandomInteger(MIN_AMOUNT_LIKES, MAX_AMOUNT_LIKES),
-  comments: Array.from({length: MAX_AMOUNT_COMMENTS_IN_OBJECT}, () => generateCommentObject())
-});
-
-// Массив объектов
-const objectsArray = Array.from({length: AMOUNT_PHOTO}, () => generateObject(generateObjectId()));
-
-// Вызов функции только для исключения ошибки в линтере
-objectsArray();
+import './data.js';
